refactor: refactor CLI
