+++ Cargo.toml
@@ -1,47 +1,2 @@
-[package]
-name = "sqllogictest"
-version = "0.5.5"
-edition = "2021"
-description = "Sqllogictest parser and runner."
-license = "MIT OR Apache-2.0"
-readme = "README.md"
-homepage = "https://github.com/risinglightdb/sqllogictest-rs"
-repository = "https://github.com/risinglightdb/sqllogictest-rs"
-keywords = ["sql", "database", "parser", "cli"]
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-bin = ["anyhow", "chrono", "console", "tokio-postgres", "env_logger", "clap", "tokio", "futures", "quick-junit", "postgres-types", "rust_decimal"]
-
-[dependencies]
-anyhow = { version = "1", optional = true }
-async-trait = "0.1"
-chrono = { version = "0.4", optional = true }
-clap = { version = "3", features = ["derive", "env"], optional = true }
-console = { version = "0.15", optional = true }
-difference = "2.0"
-env_logger = { version = "0.9", optional = true }
-futures = { version = "0.3", default-features = false, optional = true }
-futures-lite = "1"
-glob = "0.3"
-humantime = "2"
-itertools = "0.10"
-log = "0.4"
-postgres-types = { version = "0.2.3", features = ["derive","with-chrono-0_4"], optional = true }
-quick-junit = { version = "0.2", optional = true }
-rust_decimal = { version = "1.7.0", features = [ "tokio-pg" ], optional = true }
-tempfile = "3"
-thiserror = "1"
-tokio = { version = "1", features = [
-    "rt",
-    "rt-multi-thread",
-    "sync",
-    "macros",
-    "fs"
-], optional = true }
-tokio-postgres = { version = "0.7", optional = true }
-
-[[bin]]
-name = "sqllogictest"
-path = "src/main.rs"
-required-features = ["bin"]
+[workspace]
+members = ["sqllogictest", "sqllogictest-bin"]
