+++ sqllogictest-bin/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "sqllogictest-bin"
+version = "0.5.5"
+edition = "2021"
+homepage = "https://github.com/risinglightdb/sqllogictest-rs"
+keywords = ["sql", "database", "parser", "cli"]
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/risinglightdb/sqllogictest-rs"
+description = "Sqllogictest CLI."
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+anyhow = { version = "1" }
+async-trait = "0.1"
+chrono = { version = "0.4" }
+clap = { version = "3", features = ["derive", "env"] }
+console = { version = "0.15" }
+difference = "2.0"
+env_logger = { version = "0.9" }
+futures = { version = "0.3", default-features = false }
+futures-lite = "1"
+glob = "0.3"
+humantime = "2"
+itertools = "0.10"
+log = "0.4"
+postgres-types = { version = "0.2.3", features = ["derive", "with-chrono-0_4"] }
+quick-junit = { version = "0.2" }
+rust_decimal = { version = "1.7.0", features = ["tokio-pg"] }
+sqllogictest = "0.5.5"
+tempfile = "3"
+thiserror = "1"
+tokio = { version = "1", features = ["rt", "rt-multi-thread", "sync", "macros", "fs"] }
+tokio-postgres = { version = "0.7" }
