+++ sqllogictest/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "sqllogictest"
+version = "0.5.5"
+edition = "2021"
+homepage = "https://github.com/risinglightdb/sqllogictest-rs"
+keywords = ["sql", "database", "parser", "cli"]
+license = "MIT OR Apache-2.0"
+readme = "README.md"
+repository = "https://github.com/risinglightdb/sqllogictest-rs"
+description = "Sqllogictest parser and runner."
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[dependencies]
+async-trait = "0.1"
+difference = "2.0"
+futures-lite = "1"
+glob = "0.3"
+humantime = "2"
+itertools = "0.10"
+log = "0.4"
+tempfile = "3"
+thiserror = "1"
